+++ src/main/java/janggi/JanggiApplication.java
@@ -9,8 +9,8 @@ import janggi.dto.GameSummary;
 import janggi.domain.JanggiGame;
 import janggi.domain.Point;
 import janggi.dto.GameStatusInfo;
-import janggi.util.DatabaseInitializer;
-import janggi.util.JdbcConnectionManager;
+import janggi.infrastructure.DatabaseInitializer;
+import janggi.infrastructure.JdbcConnectionManager;
 import janggi.repository.JdbcGameRepository;
 import janggi.ui.InputView;
 import janggi.ui.OutputView;
