+++ src/main/java/janggi/infrastructure/DatabaseInitializer.java
@@ -1,4 +1,4 @@
-package janggi.util;
+package janggi.infrastructure;
 
 import java.io.BufferedReader;
 import java.io.IOException;
